requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
constants==0.6.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
install==1.3.4
mysql-connector-python==8.0.23
oauthlib==3.1.0
packaging==20.9
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.13.3
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
